+++ src/test/java/io/personium/core/bar/jackson/AllTests.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.personium.test.unit.cell;
+package io.personium.core.bar.jackson;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
@@ -25,7 +25,7 @@ import org.junit.runners.Suite.SuiteClasses;
  */
 @RunWith(Suite.class)
 @SuiteClasses({
-        LogTest.class
+        JSONManifestTest.class
 })
 public class AllTests {
 }
