+++ src/test/java/io/personium/core/rs/cell/AllTests.java
@@ -31,6 +31,7 @@ import io.personium.core.rs.odata.ODataMessageResourceTest;
     CellResourceTest.class,
     EventResourceTest.class,
     ODataMessageResourceTest.class,
+    LogResourceTest.class,
     TokenEndPointResourceTest.class
 })
 public class AllTests {
