+++ src/test/java/io/personium/test/RegressionTests.java
@@ -52,8 +52,8 @@ import junit.framework.TestSuite;
         io.personium.test.jersey.concurrent.AllTests.class,
         io.personium.test.performance.box.odatacol.AllTests.class,
         io.personium.test.unit.core.AllTests.class,
-        io.personium.test.unit.core.auth.AllTests.class,
-        io.personium.test.unit.core.bar.AllTests.class,
+        io.personium.core.auth.AllTests.class,
+        io.personium.core.bar.AllTests.class,
         io.personium.test.unit.core.jersey.filter.AllTests.class,
         io.personium.test.unit.core.model.impl.es.AllTests.class,
         io.personium.test.unit.core.model.impl.es.doc.AllTests.class,
